Fix markdown header syntax
The hashtag header syntax requires a space after the hashtags. Before this change, markdown renderers (ie. GitHub) were not recognizing the document's intended headers and showing the raw markdown source code. With this change, the headers are now styled, and the ugly hashtags from the markdown source code are no longer displayed to readers.